missed a listener id change
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@224091 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/jcs/auxiliary/remote/server/RemoteCacheServer.java
@@ -229,7 +229,7 @@ public class RemoteCacheServer
     public void update( ICacheElement item )
         throws IOException
     {
-        update( item, ( byte ) 0 );
+        update( item, ( long ) 0 );
     }
 
 
@@ -464,7 +464,7 @@ public class RemoteCacheServer
     public void remove( String cacheName, Serializable key )
         throws IOException
     {
-        remove( cacheName, key, ( byte ) 0 );
+        remove( cacheName, key, ( long ) 0 );
     }
